Create cpp file for automatic detection
